TODO set mpi num to int

+++ src/synapse/static/StaticData.compute.cpp
@@ -13,15 +13,15 @@ void updateStatic(Connection *connection, void *_data, real *currentE, real *cur
 	int delayLength = connection->maxDelay - connection->minDelay + 1;
 	for (int delta_t = 0; delta_t<delayLength; delta_t++) {
 		int time_idx = (time+delayLength-delta_t)%(connection->maxDelay+1);
-		int firedSize = firedTableSizes[time_idx];
+		size_t firedSize = firedTableSizes[time_idx];
 
-		for (int i=0; i<firedSize; i++) {
-			int nid = firedTable[time_idx*firedTableCap + i];
-			int startLoc = connection->pDelayStart[delta_t + nid * delayLength];
-			int synapseNum = connection->pDelayNum[delta_t + nid * delayLength];
-			for (int j=0; j<synapseNum; j++) {
+		for (size_t i=0; i<firedSize; i++) {
+		    size_t nid = firedTable[time_idx*firedTableCap + i];
+			size_t startLoc = connection->pDelayStart[delta_t + nid * delayLength];
+			size_t synapseNum = connection->pDelayNum[delta_t + nid * delayLength];
+			for (size_t j=0; j<synapseNum; j++) {
 				//int sid = connection->pSynapsesIdx[j+startLoc];
-				int sid = j+startLoc;
+				size_t sid = j+startLoc;
 				assert(sid < num);
 				real weight = data->pWeight[sid];
 				if (weight >= 0) {
